requirements.txt
asgiref==3.4.1
Django==4.0
django-appconf==1.0.5
django-cleanup==5.2.0
django-compressor==3.0
django-cors-headers==3.10.1
django-imagekit==4.1.0
django-sass-processor==1.1
django-tinymce==3.4.0
django-widget-tweaks==1.4.9
djangorestframework==3.13.1
libsass==0.21.0
mysql-connector-python==8.0.27
pilkit==2.0
Pillow==8.4.0
protobuf==3.19.1
pytz==2021.3
rcssmin==1.1.0
rjsmin==1.2.0
six==1.16.0
sqlparse==0.4.2
